+++ playground/rolldown/src/index.ts
@@ -13,3 +13,7 @@ Nam semper rutrum nulla, at tincidunt ipsum dictum eget. Nullam a ex massa. Nunc
 `;
 
 console.log(loremIpsum);
+
+import('./lazy').then(({ formatTeaser }) => {
+	console.log(formatTeaser(loremIpsum));
+});